Refactor semantic statements module into declarations and control_flow
